- Changed type of Drink.mUserRating from float to int

+++ src/com/markupartist/iglaset/provider/DrinksParser.java
@@ -120,7 +120,7 @@ class DrinksParser extends DefaultHandler {
                     mCurrentDrink.setImageUrl(Drink.ImageSize.LARGE, result);
                 }
             } else if (name.equals("user_rating") && result.length() > 0) {
-            	mCurrentDrink.setUserRating(Float.parseFloat(result));
+            	mCurrentDrink.setUserRating(Integer.parseInt(result));
             } else if (name.equals("comments")) {
             	mCurrentDrink.setCommentCount(Integer.parseInt(result));
             } else if (name.equals("ratings")) {
